+++ paint/sources/compass.c
@@ -128,9 +128,9 @@ void compass_update() {
 			if (t == _compass_hitbox_x->transform) {
 				if (mouse_started("left")) {
 					// Flip between left / right
-					_compass_compare_quat(quat_from_euler(math_pi() / (float)2, 0, math_pi() / (float)2), cq)
-					    ? viewport_set_view(-1, 0, 0, math_pi() / (float)2, 0, -math_pi() / (float)2) // Left
-					    : viewport_set_view(1, 0, 0, math_pi() / (float)2, 0, math_pi() / (float)2); // Right
+					_compass_compare_quat(quat_from_euler(math_pi() / 2.0, 0, math_pi() / 2.0), cq)
+					    ? viewport_set_view(-1, 0, 0, math_pi() / 2.0, 0, -math_pi() / 2.0) // Left
+					    : viewport_set_view(1, 0, 0, math_pi() / 2.0, 0, math_pi() / 2.0); // Right
 				}
 				gc_unroot(_compass_hovered);
 				_compass_hovered = _compass_hitbox_x;
@@ -139,9 +139,9 @@ void compass_update() {
 
 			else if (t == _compass_hitbox_y->transform) {
 				if (mouse_started("left")) {
-					_compass_compare_quat(quat_from_euler(math_pi() / (float)2, 0, math_pi()), cq)
-					    ? viewport_set_view(0, -1, 0, math_pi() / (float)2, 0, 0) // Front
-					    : viewport_set_view(0, 1, 0, math_pi() / (float)2, 0, math_pi()); // Back
+					_compass_compare_quat(quat_from_euler(math_pi() / 2.0, 0, math_pi()), cq)
+					    ? viewport_set_view(0, -1, 0, math_pi() / 2.0, 0, 0) // Front
+					    : viewport_set_view(0, 1, 0, math_pi() / 2.0, 0, math_pi()); // Back
 				}
 				gc_unroot(_compass_hovered);
 				_compass_hovered = _compass_hitbox_y;
